Additional Extensions to allow you to pass your own custom list of Personalization groups in.
Adding a couple of extra extension methods, so we can pass in our own
set of personalization groups. This enables us to get the groups from
other properties, or child items, which means we can do stuff like
customize the grid blocks (if a grid editor has a personalization group
on it, then we can call the ShowToVisitor on that block and hide / show
it accordingly.